Add package numbers to size_source for TQFN packages
These two TQFN packages generated from maxim sources refer to a PDF that
contains dozens of different package outlines. To be more specific, add
the number of the outline that was used for these footprints.

In both cases, there are multiple outlines that have exactly the same
dimensions, so just the first one from the datasheet is listed.